gitlab: adapt project and group id fields to int64 (#7670)
* gitlab: adapt project and group id fields to int64

Gitlab client v1.2.0, converts the `ID` fields to int64. This PR adapts
the functions that receive these values to int64.

Requires https://github.com/gravitational/teleport/pull/61915


* keep old version compatibility

---------


Source-commit: gravitational/teleport.e@f1a1995b8ac283257cdc6cf8c10cc06e343e65da
Source-pull-requests: gravitational/teleport.e#7670

+++ e/lib/accessgraph/gitlab/client.go
@@ -92,13 +92,13 @@ func (g gitlabClient) testTokenPermissions() error {
 // getProjectMembers returns a list of Gitlab project members
 // it uses the Gitlab API to fetch the project members
 // across all pages.
-func (g gitlabClient) getProjectMembers(projectID int) ([]*gitlab.ProjectMember, error) {
+func (g gitlabClient) getProjectMembers(projectID int64) ([]*gitlab.ProjectMember, error) {
 	opt := &gitlab.ListProjectMembersOptions{
 		ListOptions: getListOptions(),
 	}
 	var members []*gitlab.ProjectMember
 	for {
-		out, rsp, err := g.client.ProjectMembers.ListProjectMembers(projectID, opt)
+		out, rsp, err := g.client.ProjectMembers.ListProjectMembers(int(projectID), opt)
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
@@ -115,13 +115,13 @@ func (g gitlabClient) getProjectMembers(projectID int) ([]*gitlab.ProjectMember,
 // getGroupMembers returns a list of Gitlab group members
 // it uses the Gitlab API to fetch the group members
 // across all pages.
-func (g gitlabClient) getGroupMembers(groupID int) ([]*gitlab.GroupMember, error) {
+func (g gitlabClient) getGroupMembers(groupID int64) ([]*gitlab.GroupMember, error) {
 	opt := &gitlab.ListGroupMembersOptions{
 		ListOptions: getListOptions(),
 	}
 	var members []*gitlab.GroupMember
 	for {
-		out, rsp, err := g.client.Groups.ListGroupMembers(groupID, opt)
+		out, rsp, err := g.client.Groups.ListGroupMembers(int(groupID), opt)
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
